Reorganized and set up basic auto and teleop
