+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/TuningOpModes.java
@@ -40,12 +40,12 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
 import org.firstinspires.ftc.robotcore.external.navigation.UnnormalizedAngleUnit;
 import org.firstinspires.ftc.robotcore.internal.opmode.OpModeMeta;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.OTOSLocalizer;
-import org.firstinspires.ftc.teamcode.PinpointLocalizer;
-import org.firstinspires.ftc.teamcode.TankDrive;
-import org.firstinspires.ftc.teamcode.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.setup.MecanumDrive;
+import org.firstinspires.ftc.teamcode.setup.OTOSLocalizer;
+import org.firstinspires.ftc.teamcode.setup.PinpointLocalizer;
+import org.firstinspires.ftc.teamcode.setup.TankDrive;
+import org.firstinspires.ftc.teamcode.setup.ThreeDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.setup.TwoDeadWheelLocalizer;
 
 import java.util.ArrayList;
 import java.util.Arrays;
